Handle Error on client

+++ client/src/components/Banners/ListBanner.vue
@@ -11,7 +11,7 @@
     </div>
     <img :src="banner.image_url" class="img-fluid rounded shadow-sm" alt="Image Banner">
     <div class="edit position-absolute">
-      <button @click="deleteBanner(banner.id)" class="btn btn-danger btn-lg mx-2">Delete</button>
+      <button @click="deleteBanner(banner.id)" class="btn btn-outline-danger btn-lg mx-2">Delete</button>
     </div>
   </div>
 </div>
@@ -31,7 +31,19 @@ export default {
       }
     },
     deleteBanner (id) {
-      this.$store.dispatch('deleteBanner', id)
+      this.$swal({
+        title: 'Are you sure?',
+        text: 'You wont be able to revert this banner!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then(result => {
+        if (result.isConfirmed) {
+          this.$store.dispatch('deleteBanner', id)
+        }
+      })
     },
     editStatusBanner (banner) {
       const payload = {
@@ -54,8 +66,8 @@ export default {
 
 <style scoped>
 .edit {
-  bottom: 3rem;
-  right: 5rem;
+  top: 5rem;
+  right: 2rem;
 }
 img {
   object-fit: cover;
